+++ tests/wpunit/SiteTokenProviderMutationsTest.php
@@ -3,7 +3,7 @@
  * Tests Login mutation
  */
 
-class PasswordProviderMutationsTest extends \Tests\WPGraphQL\TestCase\WPGraphQLTestCase {
+class SiteTokenProviderMutationsTest extends \Tests\WPGraphQL\TestCase\WPGraphQLTestCase {
 
 	public $tester;
 	public $admin;
@@ -31,15 +31,21 @@ class PasswordProviderMutationsTest extends \Tests\WPGraphQL\TestCase\WPGraphQLT
 
 		// Set the FB provider config.
 		$this->provider_config = [
-			'name'          => 'Password',
-			'slug'          => 'password',
+			'name'          => 'Site Token',
+			'slug'          => 'siteToken',
 			'order'         => 0,
 			'isEnabled'     => true,
-			'clientOptions' => [],
-			'loginOptions'  => [],
+			'clientOptions' => [
+				'headerKey' => '',
+				'secretKey' => 'some_secret',
+			],
+			'loginOptions'  => [
+				'useAuthenticationCookie' => true,
+				'metaKey'                 => 'login',
+			],
 		];
 
-		$this->tester->set_client_config( 'password', $this->provider_config );
+		$this->tester->set_client_config( 'siteToken', $this->provider_config );
 		$this->clearSchema();
 	}
 
@@ -54,10 +60,8 @@ class PasswordProviderMutationsTest extends \Tests\WPGraphQL\TestCase\WPGraphQLT
 
 	public function login_query() : string {
 		return '
-			mutation Login( $username: String!, $password: String! ) {
-				login(
-					input: {credentials: {username: $username, password: $password }, provider: PASSWORD}
-				) {
+			mutation LoginWithSiteToken( $input: LoginInput! ) {
+				login( input: $input ) {
 					authToken
 					authTokenExpiration
 					refreshToken
@@ -85,34 +89,58 @@ class PasswordProviderMutationsTest extends \Tests\WPGraphQL\TestCase\WPGraphQLT
 	public function testLoginWithNoProvisioning() : void {
 		$query = $this->login_query();
 
-		// Test bad username.
 		$variables = [
-			'username' => 'baduser',
-			'password' => '12345',
+			'input' => [
+				'identity' => 'test_user',
+				'provider' => 'SITETOKEN',
+			],
 		];
 
-		// Test with no user to match.
+		// Test with no header key
 		$actual = $this->graphql( compact( 'query', 'variables' ) );
 
 		$this->assertArrayHasKey( 'errors', $actual );
-		// The error message changes in WP 5.7
-		$this->assertNotEmpty( $actual['errors'][0]['message'] );
+		$this->assertEquals( 'Header key for site token authentication is not defined.', $actual['errors'][0]['message'] );
 
-		// Test with bad password.
-		$variables['username'] = 'test_user';
+		// Test with no header.
+		$this->provider_config['clientOptions']['headerKey'] = 'X-My-Secret-Auth-Token';
+		$this->tester->set_client_config(
+			'siteToken',
+			$this->provider_config
+		);
+
+		$actual = $this->graphql( compact( 'query', 'variables' ) );
+
+		$this->assertArrayHasKey( 'errors', $actual );
+		$this->assertEquals( 'Missing site token in custom header.', $actual['errors'][0]['message'] );
+
+		// Test with bad header.
+		$_SERVER['HTTP_X_MY_SECRET_AUTH_TOKEN'] = 'bad_secret';
+
+		$actual = $this->graphql( compact( 'query', 'variables' ) );
+		$this->assertArrayHasKey( 'errors', $actual );
+		$this->assertEquals( 'Invalid site token.', $actual['errors'][0]['message'] );
+
+		// Test with no identity.
+		$_SERVER['HTTP_X_MY_SECRET_AUTH_TOKEN'] = 'some_secret';
+		unset( $variables['input']['identity'] );
+
+		$actual = $this->graphql( compact( 'query', 'variables' ) );
+		$this->assertArrayHasKey( 'errors', $actual );
+		$this->assertEquals( 'The SITE_TOKEN provider requires the use of the `identity` input arg.', $actual['errors'][0]['message'] );
+
+		// Test with bad identity.
+		$variables['input']['identity'] = 'bad_user';
 
 		$actual = $this->graphql( compact( 'query', 'variables' ) );
 		$this->assertArrayHasKey( 'errors', $actual );
 		$this->assertEquals( 'The user could not be logged in.', $actual['errors'][0]['message'] );
 
-		// Test with correct credentials.
-		$variables['password'] = 'test_password';
-		// Test with user already logged in.
+		// Test user already logged in.
 		wp_set_current_user( $this->test_user );
+		$variables['input']['identity'] = 'test_user';
 
-		$actual = $this->graphql( compact( 'query', 'variables' ) );
-		$this->assertArrayHasKey( 'errors', $actual );
-		$this->assertEquals( 'You are already logged in.', $actual['errors'][0]['message'] );
+		$this->assertEquals( 'The user could not be logged in.', $actual['errors'][0]['message'] );
 
 		// Test with user logged in as someone else.
 		wp_set_current_user( $this->admin );